fileクラスのファイルを分ける
