+++ 09.wc_object/lib/wc_file.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+require 'pathname'
+require_relative 'wc_text'
+
+class WcFile
+  attr_reader :name, :text
+
+  def initialize(file_name)
+    @name = file_name
+    @text = WcText.new(Pathname(@name).read)
+  end
+end
